% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uid.r
\name{uid}
\alias{uid}
\alias{hash}
\title{Generate Names}
\usage{
uid(prefix = "U")

hash(str)
}
\arguments{
\item{prefix}{A single letter (a-z or A-Z). The following defaults are
used throughout this package:
* \code{'M'} - Mutex
* \code{'Q'} - Message queue
* \code{'S'} - Semaphore
* \code{'U'} - General purpose unique id.}

\item{str}{A string (scalar character).}
}
\value{
A string (scalar character) that can be used as a mutex,
semaphore, or message queue name.
}
\description{
To ensure broad compatibility across different operating systems, names of
mutexes, semaphores, and message queues should start with a letter followed
by up to 249 alphanumeric characters. These functions generate names meeting
these requirements.\cr\cr
\itemize{
\item \code{uid()}: 13-character encoding of PID and time since epoch.
\item \code{hash()}: 11-character hash of any string (hash space = 2^64).
}
}
\details{
\code{uid()}s encode sequential millisecond intervals, beginning at the current
process's start time. If the number of requested UIDs exceeds the number of
milliseconds the process has been alive, then the process will momentarily
sleep before returning.

\code{uid()}, \code{hash()}, and \code{ps::ps_string()} all return strings of different
lengths, thereby preventing unintended collisions.
}
\examples{

    library(interprocess)
    
    uid()
    
    hash('192.168.1.123:8011')
    
}
